Modify JAR filtering command in maven-publish workflow
The earlier command in the maven-publish GitHub Action was likely causing errors due to an incorrect path in the "find" function. The command has been updated to correctly locate and copy the relevant JAR files, improving artifact upload reliability. This change will lead to an error being triggered if no matching files are found.